Fixed the main view controller so the popup works
When using a JFXPopup, the setPopupContainer must be passed a parent
node of what you pass as setSource
