+++ src/resources/fxml/Main.fxml
@@ -1,32 +1,17 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
-<?import java.lang.*?>
 <?import javafx.scene.control.*?>
-<?import javafx.scene.layout.*?>
-<?import javafx.scene.control.Button?>
 <?import javafx.scene.layout.StackPane?>
-<?import de.jensd.fx.fontawesome.AwesomeIcon?>
-<?import de.jensd.fx.fontawesome.AwesomeIconsStack?>
-<?import de.jensd.fx.fontawesome.Icon?>
-<?import javafx.scene.shape.SVGPath?>
-<?import javafx.scene.control.ProgressBar?>
-<?import javafx.scene.shape.Rectangle?>
-<?import javafx.scene.layout.GridPane?>
 <?import javafx.scene.layout.VBox?>
-<?import javafx.scene.shape.Line?>
-<?import javafx.scene.layout.Pane?>
-<?import javafx.scene.layout.HBox?>
-<?import javafx.scene.layout.AnchorPane?>
 <?import javafx.scene.control.Label?>
-<?import javafx.scene.layout.FlowPane?>
+<?import javafx.scene.layout.BorderPane?>
+<?import de.jensd.fx.fontawesome.Icon?>
 <?import com.jfoenix.controls.JFXToolbar?>
-<?import com.jfoenix.controls.JFXDrawer?>
+<?import com.jfoenix.controls.JFXListView?>
 <?import com.jfoenix.controls.JFXHamburger?>
 <?import com.jfoenix.controls.JFXRippler?>
 <?import com.jfoenix.controls.JFXButton?>
-<?import com.jfoenix.controls.JFXListView?>
 <?import com.jfoenix.controls.JFXPopup?>
-<?import javafx.scene.layout.BorderPane?>
 
 <StackPane fx:id="root" xmlns:fx="http://javafx.com/fxml/1"
   fx:controller="kabasuji.views.MainView"
